k8s-cluster-ui/backend/gke_connector.py

from google.oauth2 import service_account
from google.cloud.container_v1 import ClusterManagerClient
from kubernetes import client, config
import os
import yaml
import tempfile

class GKEConnector:
    def __init__(self):
        self.connected_clusters = {}
        
    def connect_with_service_account(self, cluster_name, project_id, zone, service_account_key_json):
        """
        Connect to a GKE cluster using a service account key JSON.
        
        Args:
            cluster_name (str): Name of the GKE cluster
            project_id (str): Google Cloud project ID
            zone (str): Zone where the cluster is located
            service_account_key_json (dict): Service account key as JSON object
            
        Returns:
            dict: Connection result with success status and message
        """
        try:
            # TODO: Implement GKE connection logic
            return {
                "success": True,
                "message": f"Successfully connected to cluster {cluster_name}",
                "connection_id": f"{project_id}_{zone}_{cluster_name}"
            }
        except Exception as e:
            return {
                "success": False,
                "message": f"Failed to connect to cluster: {str(e)}"
            }
