k8s-cluster-ui/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
kubernetes==23.6.0
google-auth==2.3.3
google-cloud-container==2.17.4
boto3==1.26.135
botocore==1.29.135
pyyaml==6.0.1
python-dotenv==0.19.2
pytest==7.3.1
gunicorn==20.1.0